[FIRRTL] Replace RHS sinks with synthetic wires
Modify the FIRRTL parser to replace right-hand-side (RHS) usages of
sinks with a synthetic wire.  A RHS sink is roughly FIRRTL's
equivalent of using an l-value as an r-value.  This is
emitted (illegally) by Chisel and supported by the Scala FIRRTL
Compiler (SFC) because it's actually useful, e.g., to read from an
output port or an instance input.

This works with a 2-pass approach in the FIRRTL parser.  (1) During
parsing, any connects, partial connects, or expressions (e.g., a
subaccess index) that have this behavior are discovered (with the help
of new utility methods) and recorded.  (2) After a module is parsed,
any recorded occurrences of RHS sinks are replaced with a synthetic
wire. The name of the synthetic wire is derived from the port or
instance port name.

The following utilities are added:

  - getKind to return a custom "kind" of "directioned" or "in-out"
  - getFlow to return the FIRRTL specification notion of "flow"
  - getPorts to find the ports associated with a value
  - getInstancePorts to find instance ports associated with a value
  - sortType to sort a FIRRTL type by bundle field name
  - commonType which returns the common parts of two types

Note: while this computes a custom "kind" and "flow", it does not
introduce the FIRRTL specification notion of flow or the SFC internal
notion of kind. The existing FIRRTL dialect type system is unmodified.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>